modification pour rendre de nouveau opérationel l'export + historisation
