+++ app/assets/stylesheets/style.css.less
@@ -221,6 +221,12 @@
   font-size: 14px;
 }
 
+.black-shadow {
+  -webkit-text-shadow: -0.5px 1px 0.10px lighten(black, 50);
+  -moz-text-shadow: -0.5px 1px 0.1px lighten(black, 50);
+  text-shadow: -0.5px 1px 0.1px lighten(black, 50);
+}
+
 #back30 {
-  margin-left: 70px;
+  margin-left: -10px;
 }
