+++ app/controllers/application_controller.rb
@@ -116,5 +116,16 @@ class ApplicationController < ActionController::Base
     relevant_phrase_text
   end
 
+  def missing_translation?(locale)
+    primary_keyword_quantity = phrase_quantity
+    current_done_translation = locale.translations.where(:needed_update => 0).count
+    primary_keyword_quantity - current_done_translation
+  end
+
+  def phrase_quantity()
+    reference_locale = Locale.primary_locale
+    primary_keyword_quantity = reference_locale.translations.count
+    primary_keyword_quantity
+  end
 
 end
