+++ app/helpers/blog_helper.rb
@@ -1,0 +1,3 @@
+module BlogHelper
+
+end
