+++ app/views/blog/index.html.haml
@@ -38,3 +38,6 @@
           = link_to(t("sign_up"), new_translator_registration_path, :class => 'btn')
     %br
     %br
+%br
+%br
+%br
