+++ config/locales/de.yml
@@ -46,3 +46,71 @@ en:
         one: " "
         other: " "
       welcome: " "
+  comments: 
+    new: 
+      new_comment: " Nouveau commentaire"
+      back: " Retour"
+  activerecord: 
+    models: 
+      user: " Utilisateur"
+    attributes: 
+      comment: 
+        email: " Courriel"
+        password: " mot de passe"
+  translations: 
+    form: 
+      phrase_id: " --"
+  errors: 
+    messages: 
+      expired: " "
+      not_found: " "
+      already_confirmed: " "
+      not_locked: " "
+      not_saved: 
+        one: " "
+        other: " "
+  devise: 
+    failure: 
+      already_authenticated: " "
+      unauthenticated: " "
+      unconfirmed: " "
+      locked: " "
+      invalid: " "
+      invalid_token: " "
+      timeout: " "
+      inactive: " "
+    sessions: 
+      signed_in: " "
+      signed_out: " "
+    passwords: 
+      send_instructions: " "
+      updated: " "
+      updated_not_active: " "
+      send_paranoid_instructions: " "
+      no_token: " "
+    confirmations: 
+      send_instructions: " "
+      send_paranoid_instructions: " "
+      confirmed: " "
+    registrations: 
+      signed_up: " "
+      signed_up_but_unconfirmed: " "
+      signed_up_but_inactive: " "
+      signed_up_but_locked: " "
+      updated: " "
+      update_needs_confirmation: " "
+      destroyed: " "
+    unlocks: 
+      send_instructions: " "
+      unlocked: " "
+      send_paranoid_instructions: " "
+    omniauth_callbacks: 
+      success: " "
+      failure: " "
+    mailer: 
+      confirmation_instructions: 
+        subject: " "
+      reset_password_instructions: 
+        subject: " "
+      unlock_instructions: 
+        subject: " "
