+++ config/locales/de.yml.old
@@ -1,48 +1,3 @@
-en: 
+en:
   home: " Houssen"
   article: " Articlen"
-  team: " Speal"
-  about: " lol"
-  logout: " déconnexionen"
-  login: " Connexion"
-  sign_up: " S'enregistrer"
-  articles: 
-    index: 
-      show: " Plus d'information »"
-      subject: " Sujet"
-      cattegory: " Category"
-      posted: " Posté le"
-      new_article: " Déposé un nouvel article"
-      article_list: " Liste d'article"
-      published_article_quantity: 
-        zero: " Il n'a aucun article publié dans la base de données"
-        one: " Il y a un article publié dans la base de donnée"
-        other: " Il y a %{count} articles publiés dans la base de données"
-      article_quantity: 
-        zero: " Il n'y a pas d'article dans la base de données"
-        one: " Il y a un article dans la base de donnée"
-        other: " Il y a %{count} articles dans la base de donnée"
-    show: 
-      comment: " Commenter"
-      subject: " Sujet"
-      from: " De"
-      category: " Catégorie"
-      content: " Contenu"
-      comments: " Commentaires"
-      back: " Retour"
-      comment_quantity: 
-        zero: " Il y a rien"
-        one: " "
-        other: " "
-    new: 
-      new_article: " Post un nouvel article"
-  blog: 
-    index: 
-      bonjour: " Salut"
-      welcome_paragraphe: " Vous êtes sur le Blog de Djengo"
-      welcome_paragraphe_2: " "
-      require: 
-        zero: " "
-        one: " "
-        other: " "
-      welcome: " "
