+++ config/locales/en.yml
@@ -2,12 +2,6 @@
 # See https://github.com/svenfuchs/rails-i18n/tree/master/rails%2Flocale for starting points.
 
 en:
-  todo:
-    - Sun
-    - Sun
-    - Sun
-    - Sun
-    - Sun
   date:
     abbr_day_names:
     - Sun
